requirements.txt
assembly==1.3.0
interaction==1.3
material==0.1
matplotlib==3.8.4
mesh==1.0.0a1
numpy==2.1.0
pydicom==2.4.4
PyYAML==6.0.2
section==2.0
sketch==0.5.2
step==0.1.0
visualization==1.0.0
xyPlot==0.2.0
